Moved relevant game logic to MinesweeperGameManager
